+++ nari/types/event/instance.py
@@ -4,9 +4,16 @@ from enum import Enum, auto
 
 from nari.types.event import Event
 
+
 class Fade(Enum):
     In = auto()
     Out = auto()
+
+
+class BarrierState(Enum):
+    down = auto()
+    up = auto()
+
 
 class InstanceComplete(Event): # pylint: disable=too-few-public-methods
     """Represents an instance being complete"""
@@ -47,6 +54,21 @@ class InstanceFade(Event): # pylint: disable=too-few-public-methods
 
     def __repr__(self):
         return '<InstanceFade>'
+
+
+class BarrierToggle(Event): # pylint: disable=too-few-public-methods
+    """Represents a barrier changing state"""
+    def __init__(self, *,
+                 timestamp: datetime,
+                 instance_id: int,
+                 state: BarrierState,
+                ):
+        super().__init__(timestamp)
+        self.instance_id = instance_id
+        self.state = state
+
+    def __repr__(self):
+        return '<BarrierToggle>'
 
 
 class InstanceInit(Event): # pylint: disable=too-few-public-methods
