--- nari/types/event/statusremove.py
@@ -1,27 +1,0 @@
-"""Stuff for removal of statuses"""
-from datetime import datetime
-
-from nari.types.event import Event
-from nari.types.actor import Actor
-from nari.types.status import Status
-
-
-class StatusRemove(Event): # pylint: disable=too-few-public-methods
-    """Event represents the application of a status"""
-    def __init__(self, *,
-                 timestamp: datetime,
-                 status: Status,
-                 duration: float,
-                 source_actor: Actor,
-                 target_actor: Actor,
-                 params: int
-                ):
-        super().__init__(timestamp)
-        self.status = status
-        self.duration = duration
-        self.source_actor = source_actor
-        self.target_actor = target_actor
-        self.params = params
-
-    def __repr__(self):
-        return '<StatusRemove>'
