+++ nari/types/statuseffect.py
@@ -1,6 +1,9 @@
-from typing import NamedTuple, List
+"""Types revolving around status effects"""
+
+from typing import NamedTuple
 
 class StatusEffect(NamedTuple):
+    """Container for parameters for status effects"""
     status_id: int
     status_params: bytes
     duration: float
